requirements.txt
gymnasium==0.29.0
shimmy<2.0
dm_control
mujoco<3.0,>=2.3.3
numpy >= 1.20.2
jax >= 0.4.16
flax >= 0.7.2
ml_collections >= 0.1.0
tqdm >= 4.60.0
optax >= 0.1.3
absl-py >= 0.12.0
scipy >= 1.6.0
wandb >= 0.16.2
moviepy
imageio
distrax